+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
   ## 0.4.12
   - Upgrade minimum VS Code version to 1.125.
+    - Fix: `standardrb` cannot use bundler when `formatto.standardrbFormattingMode` is set to `tmpFile`. 
 
   ## 0.4.11
   - Default `formatto.excludePatterns` to `[vendor/**]`.